+++ test/instaparseclr/viz_test.clj
@@ -1,31 +1,31 @@
-(ns instaparse.viz-test
-  (:require instaparse.core)
-  (:use instaparse.viz))
+(ns instaparseclr.viz-test
+  (:require instaparseclr.core)
+  (:use instaparseclr.viz))
 
 (def make-tree-e
      "simple tree parser"
-     (instaparse.core/parser "tree: node* 
+     (instaparseclr.core/parser "tree: node* 
               node: leaf | <'('> node (<'('> node <')'>)* node* <')'> 
               leaf: #'a+'
               " :output-format :enlive))
 
 (def make-tree-h
      "simple tree parser"
-     (instaparse.core/parser "tree: node* 
+     (instaparseclr.core/parser "tree: node* 
               node: leaf | <'('> node (<'('> node <')'>)* node* <')'> 
               leaf: #'a+'
               " :output-format :hiccup))
 
 (def make-tree-se
      "simple tree parser"
-     (instaparse.core/parser "<tree>: node* 
+     (instaparseclr.core/parser "<tree>: node* 
               node: leaf | <'('> node (<'('> node <')'>)* node* <')'> 
               leaf: #'a+'
               " :output-format :enlive))
 
 (def make-tree-sh
      "simple tree parser"
-     (instaparse.core/parser "<tree>: node* 
+     (instaparseclr.core/parser "<tree>: node* 
               node: leaf | <'('> node (<'('> node <')'>)* node* <')'> 
               leaf: #'a+'
               " :output-format :hiccup))